pyproject.toml
[tool.poetry]
name = "pub-ijar-learning_cascades"
version = "1.0.0"
description = ""
authors = ["Jonas Hanselle <jonas.hanselle@ifi.lmu.de>", "Stefan Heid <stefan.heid@ifi.lmu.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
scikit-learn = "^1.3.0"
numpy = "^1.25.2"
scipy = "^1.11.2"
joblib = "^1.3.2"
pandas = "^2.1.0"
sortedcontainers = "^2.4.0"
scikit-psl = "0.6.0"
sklearn-riskslim-slim = { git = "git@github.com:stheid/sklearn-riskslim-slim.git", rev = "047f924"}
jupyterlab = "^4.0.7"
tqdm = "^4.66.1"
matplotlib = "^3.8.2"
seaborn = "^0.13.0"
latex = "^0.7.0"
ucimlrepo = "^0.0.3"
networkx = "^3.2.1"

[build-system]
requires = ["poetry-core>=1.6.0"]
build-backend = "poetry.core.masonry.api"
